jenkins/custom_tools/DevopsJsonConv/src/main/java/com/infosys/json/SoapUIFailure.java
package com.infosys.json;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

public class SoapUIFailure {

	@SerializedName("type")
	@Expose
	private String type;
	@SerializedName("message")
	@Expose
	private String message;
	public String getType() {
		return type;
	}
	public void setType(String type) {
		this.type = type;
	}
	public String getMessage() {
		return message;
	}
	public void setMessage(String message) {
		this.message = message;
	}
	
	
}
